preflight: initial module commit
This is based on an early version of Stephen Dade's checklist module,
and provides a digital version of the pre-flight process used by the
NPS team. Features include verification of parameters, confirmation
text boxes for per-flight settings, and single-click access for a
flight technician (operating a separate instance of MAVProxy from the
GCS operator) to perform a final run-up (pitot calibration, INS/IMU
checks, throttle run-up, etc).

Note that this module references parameters, etc that are specific
to NPS (ACS).