preflight: added airspeed sensor check
Airspeed sensor might be miscalibrated, in which case it will
misreport airspeed during flight and possibly lead to a stall.
The symptom of miscalibration is "flatline" (zero) airspeed
reports in the VFR_HUD mavlink message.

This new check maintains a history of the last N airspeed values,
and sets the current airspeed reading green or red based not only
on absolute value but also on whether there is *any* variance in
the readings (no variance = bad).